requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
openai
anthropic
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0